Remove unnecessary explicit unboxing

+++ com.ibm.wala.cast.java.test.data/src/AnonymousClass.java
@@ -25,7 +25,7 @@ public class AnonymousClass {
       public int getValue() { return value; }
 	    
       @Override
-      public int getValueBase() { return value - base.intValue(); }
+      public int getValueBase() { return value - base; }
 	};
 
 	System.out.println(f.getValue());
@@ -51,7 +51,7 @@ public class AnonymousClass {
 	    
       @Override
       public int getValueBase() { 
-	      return y + getValue() - base.intValue(); 
+	      return y + getValue() - base;
 	    }
 	}
